pyproject.toml
[tool.poetry]
name = "roombasdk"
version = "1.7.7"
description = "Python program and library to control Wi-Fi enabled iRobot Roombas"
authors = ["Philipp Schmitt <philipp@schmitt.co>", "Jack Simbach <jack.simbach@gmail.com"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/simbaja/roombapy"
classifiers = [
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Networking"
]
packages = [
    {include = "roombapy"}
]
exclude = ["tests/*", "samples/*"]

[tool.poetry.scripts]
roomba-connect = "roombapy.entry_points:connect"
roomba-discovery = "roombapy.entry_points:discovery"
roomba-password = "roombapy.entry_points:password"

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
paho-mqtt = "^1.5.1"
pillow = ">=8.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
black = "^21.9b0"
isort = "^5.9.3"
flake8 = "^3.9.2"
amqtt = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
